+++ app/Actions/Fortify/UpdateUserProfileInformation.php
@@ -19,9 +19,14 @@ class UpdateUserProfileInformation implements UpdatesUserProfileInformation
     {
         Validator::make($input, [
             'name' => ['required', 'string', 'max:255'],
+            'username' => ['string', 'max:144'],
             'email' => ['required', 'email', 'max:255', Rule::unique('users')->ignore($user->id)],
             'photo' => ['nullable', 'mimes:jpg,jpeg,png', 'max:1024'],
         ])->validateWithBag('updateProfileInformation');
+
+        if (isset($input['username'])) {
+            $user->username = $input['username'];
+        }
 
         if (isset($input['photo'])) {
             $user->updateProfilePhoto($input['photo']);